refactor(web): split contact page sub-components
Move the inline editors (communication, custom fields, owners, notes,
editable rows) and shared rows/icons out of the contact route file
into focused modules under -components/. Route stays as the entry +
JSX layout; component shapes and behaviour unchanged.
