+++ apps/web/src/routes/_protected.contacts/-components/email-link.tsx
@@ -1,0 +1,8 @@
+export const EmailLink = ({ address }: { address: string }) => (
+  <a
+    className="hover:text-foreground min-w-0 break-all hover:underline"
+    href={`mailto:${address}`}
+  >
+    {address}
+  </a>
+);
